requirements.txt
langchain
langchain-core
langchain-huggingface
python-dotenv
scikit-learn
numpy